// angular-src/src/app/model/user-activity.ts
/**
 * A model for an user activity
 */
export class UserActivity {
    userId: string;
    timestamp: string;
    log: string;
    

    constructor(userId: string, timestamp: string, log: string){
        // this.userId = userId;
        // this.timestamp = timestamp;
        // this.log = log;
    }
}